Consolidate all UserService tests into UserServiceImplTest
